hw-3-unit-tests/src/com/github/basetask/arrays/ArraysTask9SelectSort.java
package com.github.basetask.arrays;

import java.util.Arrays;
import java.util.Scanner;

public class ArraysTask9SelectSort {
    public static void main(String[] args) {
        Scanner scan = new Scanner(System.in);
        System.out.println("Enter array length: ");
        int size = scan.nextInt();
        int[] array = new int[size];
        System.out.println("Insert array elements: ");
        for(int i = 0; i < size; i++){
            array[i] = scan.nextInt();
        }
        int[] sortedArray = selectSort(array);
        System.out.println("An array sorted with Selection Sort: " + Arrays.toString(sortedArray));
    }
   public static int[] selectSort(int[] array){
        int size = array.length;
        for(int i = 0; i < size; i++){
            for(int j = i+1; j < size; j++){
                if(array[j] < array[i]){
                    int temp = array[j];
                    array[j] = array[i];
                    array[i] = temp;
                }
            }
        }
        return array;
    }
}
